source/lambda/ingestion/requirements.txt
boto3>=1.34.0
pypdf==4.2.0
python-docx==1.1.2
